Remove relative paths; Add new 404 page; Reorganize assets
(1) Remove relative paths - I went through all of the files and removed
all instances of "../.." and so on. Now, every file path begins from
the root ("/"). Sadly this does not work locally, but it does work
remotely.

(2) Add new 404 page - I finally copied over the HISS 404 terminal emulator
page to my personal site. I didn't want to part with my old 404 page, so I
found a way to forward the user randomly to one of the pages.

(3) Reorganize assets - I renamed landing_page_assets/ to assets/ and made
some somedirectories in there for to be more organized. Arguably was more
trouble than it's worth, and I might revert this later.