(blog): simplify blog structure + general cleanup
The old blog format was ambitious in that I was trying to
get transclusion to work for blog posts. But honestly, I never
really used that feature, I like the 'archive' blog format much
better. My implementation also had multiple non-compliant HTML
things happening, some on purpose and some not. In general, it was
complex enough to make me not want to post stuff.

This new setup is very straightforward, every title is just a link.
It also comes with the added benefit of not forcing users to have
JavaScript enabled in order for stuff to make sense.