(blog): change footer to say 'html compiler'
When I originally wrote the footer (and phd, the markdown compiler) I didn't
understand what a parser was, but I refused to accept that what I'd made was
a compiler (after all, compilers are cool and hard to build and stuff). I
also considered that maybe it was just a "converter", but that was a bit too
generic. Later I thought that maybe it was a transpiler, but the distinction
between that and a compiler is kind of arbitrary. Anyways, it's a compiler, and
a damn shitty one at that :-).