(albums): fix page styles for Firefox
This page's styles use the non-standard `zoom` element which is unsupported
in Firefox. As a result the page looked super wonky on Firefox, which was
making me sad. With the workaround presented here, the page's rendering on
Firefox has basically full parity with Safari and Chrome.

Additionally, towards the goal of making the page render as intended in all
browsers, I put in a fix for the audio element. In Firefox it was causing
all of the album "cards" to take up a constant width, which was overly-wide
for most of them (as opposed to fitting to the width of the content). I was
accidentally relying on some weird overflow behavior to make the element
break to the next line, so I changed it to be explicit and that fixed the
problem.